(Build) Fix checks for gles20, glyphcache install folder
Tizen Common does not have "gles20", but only "glesv2" and
"egl" in all repos. We have a macro to replace it in the
specfile, but not the autoconf files. Fixing.

The glyphcache install folder seems to have changed in
latest releases ("/opt/usr" to "/usr"). Fixing.

Use the %configure macro instead of "./configure". Simplify
the specfile by using a conditional variable.

Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>

Change-Id: Icc88ee08b2f373be92e1a9be0f85656da032f99e