Stopped log messages from crowding a single line in desktop builds
[problem]
Some of our error log items are not printed on their own lines in
desktop builds.
An example is this item, which piles up on a single line when itemview
and other demos are exited by closing the demo window:
  ERROR: DALI: Dali::Internal::GlyphStatusContainer::~GlyphStatusContainer() glyph still referenced (leaking)

[cause]
ERROR_LOG output requires an explicit terminating newline in desktop
builds but not in target ones.
Some of our logging has terminating newlines at the calling site and
some not.

[solution]
Trim newlines and other whitespace from the end of log messages and
then add a single newline back on.

Change-Id: Ia6a227db2be3929f7ee85ff7255b4526344d0c42
Signed-off-by: Andrew Cox <andrew.cox@partner.samsung.com>