Moved NotificationTrigger from update thread to update-render-sync
Need to prevent event triggers from being sent when the system is
shutting down. The update thread gets this information after triggering
the event thread to wake up.

Instead, by moving the notification trigger to the UpdateRenderSynchronisation
class, we can ensure that the trigger is only sent whilst mRunning is true;
i.e. that the trigger is not called after the Adaptor destructor has been called.

The Trigger() method is still expected not to block.

Change-Id: Id0394fa0bc32f2dbf03f1edeae1a2b46ad7cb90b