TEMP CHANGE for android adaptor
This comprises the current state of the android adaptor for Dali.
The build scripts under build/android are intended for building all 3 repos
and should really be at a level higher than this repo's root dir.
(Intention is to split up the build into the respective repos, as this generates
a static library; which is ideal for splitting into 3)

It can be used with simple examples in dali demo, though it only has support for
png images at the present time (adding jpeg support is on-going)

Change-Id: I6826a038b3a36939100766bdc7243bdcf9cfd4c1
Signed-off-by: David Steele <david.steele@partner.samsung.com>