Removed singletons from the type registry
The type registry creation order is dependent on library load order, so it's possible for dali toolkit singletons to be created before adaptor singletons.

Have changed all adaptor singletons to register with the singleton
service in the Get() method, and removed unnecessary type registry
entries for them. Thus, they will be created when first required,
rather than when the type registry tries to create them.

Change-Id: I36281d86f5c769a05feb6c074b55375ddea9359b
Signed-off-by: David Steele <david.steele@samsung.com>